renamed skeleton.c to input_buffer.c + changed makefile correspondingly
