+++ Makefile
@@ -21,7 +21,7 @@ CLIENT_SRCS = client.c
 CLIENT_BIN = client
 
 # Simple Database Management System
-SDBMS_SRCS = database/skeleton.c database/sdbms.c
+SDBMS_SRCS = database/input_buffer.c database/sdbms.c
 SDBMS_BIN = sdbms
 
 all: $(THREADPOOL_BIN) $(WEBSERVER_BIN) $(SSH_SERVER_BIN) $(SSH_CLIENT_BIN) $(CLIENT_BIN) $(SDBMS_BIN)
